@@ -1,9 +1,10 @@
 package avida.ican.Farzin.View.Dialog;
 
 import android.app.Activity;
-import android.support.annotation.Nullable;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.view.ViewPager;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.viewpager.widget.ViewPager;
 import android.view.Gravity;
 import android.view.View;
 import android.view.ViewGroup;
@@ -121,7 +122,7 @@ public class DialogZanjireMadrak {
     }
 
     private void clearFragment() {
-        android.support.v4.app.FragmentTransaction transaction = mfragmentManager.beginTransaction();
+        FragmentTransaction transaction = mfragmentManager.beginTransaction();
         transaction.remove(fragmentZanjireMadrak);
         transaction.commitAllowingStateLoss();
     }
