@@ -5,9 +5,9 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.design.widget.BottomNavigationView;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.Menu;
@@ -362,11 +362,11 @@ public class FarzinActivityMain extends BaseNavigationDrawerActivity implements 
     }
 
     public void selectMessageFragment(boolean isFilter) {
-        if (!filtering) {
-            if (staticbottomNavigation != null) {
-                this.isFilter = isFilter;
-                staticbottomNavigation.setCurrentItem(2);
-            }
+        if (staticbottomNavigation != null) {
+            this.isFilter = isFilter;
+            staticbottomNavigation.setCurrentItem(2);
+        }
+        if (isFilter) {
             filterReceiveMessage();
         }
 
