@@ -3,10 +3,10 @@ package avida.ican.Farzin.View.Fragment.Cartable;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+import androidx.annotation.Nullable;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
@@ -17,22 +17,14 @@ import android.widget.TextView;
 import java.util.ArrayList;
 import java.util.List;
 
-import avida.ican.Farzin.Model.Enum.ZanjireMadrakFileTypeEnum;
 import avida.ican.Farzin.Model.Structure.Database.Cartable.StructureCartableHistoryDB;
-import avida.ican.Farzin.Model.Structure.Database.Cartable.StructureZanjireMadrakFileDB;
 import avida.ican.Farzin.Model.Structure.Response.Cartable.StructureHistoryListRES;
 import avida.ican.Farzin.Model.Structure.Response.Cartable.StructureNodeRES;
 import avida.ican.Farzin.Presenter.Cartable.FarzinHistoryListPresenter;
-import avida.ican.Farzin.Presenter.Cartable.FarzinZanjireMadrakPresenter;
-import avida.ican.Farzin.View.Adapter.AdapteZanjireMadrak;
 import avida.ican.Farzin.View.Adapter.AdapterCartableHistory;
-import avida.ican.Farzin.View.Interface.Cartable.ListenerAdapterZanjireMadrak;
 import avida.ican.Farzin.View.Interface.Cartable.ListenerGraf;
-import avida.ican.Farzin.View.Interface.Cartable.ListenerZanjireMadrak;
-import avida.ican.Farzin.View.Interface.ListenerFile;
 import avida.ican.Ican.App;
 import avida.ican.Ican.BaseFragment;
-import avida.ican.Ican.Model.Structure.StructureAttach;
 import avida.ican.Ican.View.Custom.CustomFunction;
 import avida.ican.Ican.View.Custom.GridLayoutManagerWithSmoothScroller;
 import avida.ican.Ican.View.Enum.NetworkStatus;
