@@ -2,7 +2,9 @@ package avida.ican.Ican.View.Adapter;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
