@@ -3,7 +3,7 @@ package avida.ican.Ican.View.Custom;
 
 import android.annotation.SuppressLint;
 import android.os.AsyncTask;
-import android.support.design.widget.Snackbar;
+import com.google.android.material.snackbar.Snackbar;
 import android.view.View;
 import android.widget.Toast;
 
